Parametric bootstrap

Metida.bootstrap

Metida.bootstrap — Function

bootstrap(lmm::LMM; double = true, n = 100, varn = n, verbose = true, init = lmm.result.theta, rng = default_rng())

Parametric bootstrap.

Warning

Experimental: API not stable, results not validated

double - use double approach (default - true);

n - number of bootstrap samples for coefficient estimtion;

varn - number of bottstrap samples for varianvce estimation;

verbose - show progress bar;

init - initial values for lmm;

rng - random number generator.

Parametric bootstrap based on generating random responce vector from known distribution, that given from fitted LMM model. For one-stage bootstrap variance parameters and coefficients simulated in one step. For double bootstrap (two-tage) variance parameters simulated first, than used for simulating coefficients on stage two.

lmm = Metida.LMM(@formula(var~sequence+period+formulation), df0m;
random = Metida.VarEffect(Metida.@covstr(formulation|subject), Metida.CSH),
)
Metida.fit!(lmm)
bt = Metida.bootstrap(lmm; n = 1000, varn = 1000, double = true, rng = MersenneTwister(1234))
confint(bt)

See also: confint, Metida.miboot