Installation

import Pkg; Pkg.add("Metida")

Simple example

LMM Metida.@covstr Metida.VarEffect fit!

Step 1: Load data

Load provided data with CSV and DataFrames:

using CSV, DataFrames
df = CSV.File(joinpath(dirname(pathof(Metida)), "..", "test", "csv", "df0.csv")) |> DataFrame

Note

Check that all categorical variables is categorical.

categorical!(df, :subject);
categorical!(df, :period);
categorical!(df, :sequence);
categorical!(df, :formulation);

Step 2: Make model

lmm = LMM(@formula(var~sequence+period+formulation), df;
random = VarEffect(@covstr(formulation|subject), CSH),
repeated = VarEffect(@covstr(formulation|subject), DIAG))

Step 3: Fit

fit!(lmm)

Model construction

Metida.LMM — Type

LMM(model, data; contrasts=Dict{Symbol,Any}(),  random::Union{Nothing, VarEffect, Vector{VarEffect}} = nothing, repeated::Union{Nothing, VarEffect} = nothing)

Make Linear-Mixed Model object.

model: is a fixed-effect model (@formula)

data: tabular data

contrasts: contrasts for fixed factors

random: vector of random effects or single random effect

repeated: is a repeated effect (only one)

model - example: @formula(var ~ sequence + period + formulation)

random - effects can be specified like this: VarEffect(@covstr(formulation|subject), CSH). @covstr is a effect model: @covstr(formulation|subject). CSH is a CovarianceType structure. Premade constants: SI, DIAG, AR, ARH, CS, CSH, ARMA. If not specified only repeated used.

repeated - can be specified like random effect. If not specified VarEffect(@covstr(1|1), SI) used. If no repeated effects specified vector of ones used.

Random/repeated model

Metida.@covstr — Macro

@covstr(ex)

Macros for random/repeated effect model.

Example: @covstr(factor|subject)

Random/repeated effect construction

Metida.VarEffect — Type

VarEffect(formula, covtype::T, coding) where T <: AbstractCovarianceType

VarEffect(formula, covtype::T; coding = nothing) where T <: AbstractCovarianceType

VarEffect(formula; coding = nothing)

Random/repeated effect.

formula from @covstr(ex) macros.

covtype - covariance type (SI, DIAG, CS, CSH, AR, ARH, ARMA)

Fitting

StatsBase.fit! — Function

fit!(lmm::LMM{T};
solver::Symbol = :default,
verbose = :auto,
varlinkf = :exp,
rholinkf = :sigm,
aifirst::Bool = false,
g_tol::Float64 = 1e-12,
x_tol::Float64 = 1e-12,
f_tol::Float64 = 1e-12,
hcalck::Bool   = true,
init = nothing,
io::IO = stdout,
) where T

Fit LMM model.

solver - :default / :nlopt / :cuda

verbose - :auto / 1 / 2 / 3

varlinkf - not implemented

rholinkf - :sigm / :atan

aifirst - first iteration with AI-like method

g_tol - absolute tolerance in the gradient

x_tol - absolute tolerance of theta vector

f_tol - absolute tolerance in changes of the REML

hcalck - calculate REML Hessian

init - initial theta values

io - uotput IO

solver - :default solving with Optim.jl, for :nlopt and :cuda MetidaNLopt.jl and MetidaCu.jl should be installed.

verbose - 1 - only log, 2 - log and print, 3 - print only errors, other log, 0 (or any other value) - no logging.

Custom structure

Step 1: make CustomCovarianceStruct

Metida.CustomCovarianceStruct — Type

CustomCovarianceStruct(nparamf::Function, xmat!::Function)

nparamf - function type (t, q) -> (a, b)

where: t - size(z, 2) - number of levels for effect (number of columns of individual z matriz); q - number of factors in the effect model; a - number of variance parameters; b - number of ρ parameters;

Example: (t, q) -> (t, 1) for CSH structure; (t, q) -> (1, 1) for AR, ets.

Tuple{Int, Int} should be returned.

xmat! - construction function

G matrix function should update mx, where mx is zero matrix, p - parameter of CovarianceType structure, example:

function gmat_diag!(mx, θ::Vector{T}, p) where T for i = 1:size(mx, 1) mx[i, i] = θ[i] ^ 2 end nothing end

R matrix function should add R part to mx, rz - is repeated effect matrix, example:

function rmatp_diag!(mx, θ::Vector{T}, rz, p) where T for i = 1:size(mx, 1) for c = 1:length(θ) mx[i, i] += rz[i, c] * θ[c] * θ[c] end end nothing end

Step 2: make CustomCovarianceType

Metida.CustomCovarianceType — Function

CustomCovarianceType(ccs::CustomCovarianceStruct)

Make custom covariance type with CustomCovarianceStruct.

Step 3 Fit your model

using Metida, StatsPlots, CSV, DataFrames, MixedModels; #hide

rds = CSV.File(joinpath(dirname(pathof(Metida)), "..", "test", "csv",  "df0.csv"); types = [String, String, String, String, Float64, Float64]) |> DataFrame

#Make struct for G and R matrix
ccsg = CustomCovarianceStruct((q,p) -> (q, 1), Metida.gmat_csh!)
ccsr = CustomCovarianceStruct((q,p) -> (q, 0), Metida.rmatp_diag!)

#Make type struct
CCTG = CustomCovarianceType(ccsg)
CCTR = CustomCovarianceType(ccsr)

#Make model
lmm = LMM(@formula(var~sequence+period+formulation), df0;
random = VarEffect(@covstr(formulation|subject), CCTG),
repeated = VarEffect(@covstr(formulation|subject), CCTR),
)

#Fit model
fit!(lmm)